+++ stackslib/src/core/mempool.rs
@@ -569,7 +569,7 @@ pub struct MemPoolWalkSettings {
 impl Default for MemPoolWalkSettings {
     fn default() -> Self {
         MemPoolWalkSettings {
-            strategy: MemPoolWalkStrategy::GlobalFeeRate,
+            strategy: MemPoolWalkStrategy::NextNonceWithHighestFeeRate,
             max_walk_time_ms: u64::MAX,
             consider_no_estimate_tx_prob: 5,
             nonce_cache_size: 1024 * 1024,
@@ -583,7 +583,7 @@ impl Default for MemPoolWalkSettings {
 impl MemPoolWalkSettings {
     pub fn zero() -> MemPoolWalkSettings {
         MemPoolWalkSettings {
-            strategy: MemPoolWalkStrategy::GlobalFeeRate,
+            strategy: MemPoolWalkStrategy::NextNonceWithHighestFeeRate,
             max_walk_time_ms: u64::MAX,
             consider_no_estimate_tx_prob: 5,
             nonce_cache_size: 1024 * 1024,
@@ -1435,11 +1435,22 @@ impl MemPoolDB {
 
     #[cfg_attr(test, mutants::skip)]
     pub fn reset_mempool_caches(&mut self) -> Result<(), db_error> {
-        debug!("reset nonce cache");
-        // Delete all rows from the nonces table
-        self.db.execute("DELETE FROM nonces", NO_PARAMS)?;
-        // Also delete all rows from the considered_txs table
+        self.reset_nonce_cache()?;
+        self.reset_considered_txs_cache()?;
+        Ok(())
+    }
+
+    #[cfg_attr(test, mutants::skip)]
+    pub fn reset_considered_txs_cache(&mut self) -> Result<(), db_error> {
+        debug!("reset considered txs cache");
         self.db.execute("DELETE FROM considered_txs", NO_PARAMS)?;
+        Ok(())
+    }
+
+    #[cfg_attr(test, mutants::skip)]
+    pub fn reset_nonce_cache(&mut self) -> Result<(), db_error> {
+        debug!("reset nonce cache");
+        self.db.execute("DELETE FROM nonces", NO_PARAMS)?;
         Ok(())
     }